@@ -1,6 +1,6 @@
 {
     "name": "Pangalinker",
-    "version": "1.1.0",
+    "version": "1.2.0",
     "private": true,
     "description": "Pangalink test application",
     "scripts": {
@@ -12,44 +12,45 @@
         "url": "git://github.com/andris9/pangalink.net.git"
     },
     "author": "Andris Reinman",
-    "license": "EUPL-1.1",
+    "license": "MIT",
     "dependencies": {
-        "bcryptjs": "^2.4.3",
-        "compression": "^1.7.1",
-        "config": "^1.26.2",
-        "connect-flash": "^0.1.1",
-        "connect-mongo": "^1.3.2",
-        "cookie-parser": "^1.4.3",
-        "csurf": "^1.9.0",
-        "diacritics": "^1.3.0",
-        "ejs": "^2.5.7",
-        "encoding": "^0.1.12",
-        "express": "^4.16.1",
-        "express-session": "^1.15.6",
-        "fetch": "^1.1.0",
-        "he": "^1.1.1",
-        "highlight.js": "^9.12.0",
-        "iban": "0.0.8",
-        "joi": "^11.1.1",
-        "marked": "^0.3.6",
-        "moment": "^2.18.1",
-        "mongodb": "^2.2.31",
-        "morgan": "^1.9.0",
-        "nodemailer": "^4.1.2",
-        "npmlog": "^4.1.2",
-        "on-headers": "^1.0.1",
-        "passport": "^0.4.0",
-        "passport-local": "^1.0.0",
-        "pem": "^1.12.0",
-        "random-string": "^0.2.0",
-        "st": "^1.2.0",
-        "zip-stream": "^1.2.0"
+        "bcryptjs": "2.4.3",
+        "compression": "1.7.3",
+        "config": "2.0.1",
+        "connect-flash": "0.1.1",
+        "connect-mongo": "2.0.1",
+        "cookie-parser": "1.4.3",
+        "csurf": "1.9.0",
+        "diacritics": "1.3.0",
+        "ejs": "2.6.1",
+        "encoding": "0.1.12",
+        "express": "4.16.3",
+        "express-session": "1.15.6",
+        "fetch": "1.1.0",
+        "he": "1.1.1",
+        "highlight.js": "9.12.0",
+        "iban": "0.0.11",
+        "joi": "13.6.0",
+        "marked": "0.5.0",
+        "moment": "2.22.2",
+        "mongodb": "3.1.4",
+        "morgan": "1.9.0",
+        "nodemailer": "4.6.8",
+        "npmlog": "4.1.2",
+        "on-headers": "1.0.1",
+        "passport": "0.4.0",
+        "passport-local": "1.0.0",
+        "pem": "1.12.5",
+        "random-string": "0.2.0",
+        "st": "1.2.2",
+        "zip-stream": "2.0.1"
     },
     "devDependencies": {
-        "eslint": "^4.8.0",
-        "eslint-config-nodemailer": "^1.2.0",
-        "grunt": "^1.0.1",
-        "grunt-cli": "^1.2.0",
-        "grunt-eslint": "^20.1.0"
+        "eslint": "5.5.0",
+        "eslint-config-nodemailer": "1.2.0",
+        "eslint-config-prettier": "3.0.1",
+        "grunt": "1.0.3",
+        "grunt-cli": "1.3.1",
+        "grunt-eslint": "21.0.0"
     }
 }
