@@ -210,12 +210,12 @@ IPAYServlet.prototype.validateSignature = function(callback) {
                 success: !!success,
                 errors: !success
                     ? [
-                        {
-                            field: 'mac',
-                            error: util.format('Allkirja parameetri %s valideerimine ebaõnnestus.', 'mac'),
-                            download: true
-                        }
-                    ]
+                          {
+                              field: 'mac',
+                              error: util.format('Allkirja parameetri %s valideerimine ebaõnnestus.', 'mac'),
+                              download: true
+                          }
+                      ]
                     : false,
                 warnings: false
             });
@@ -579,11 +579,11 @@ IPAYServletValidator.prototype.validate_mac = function() {
         return util.format('Allkirja parameeter %s peab olema HEX formaadis', 'mac');
     }
 
-    if (new Buffer(value, 'hex').length % 128) {
+    if (Buffer.from(value, 'hex').length % 128) {
         return util.format(
             'Allkirja parameeter %s on vale pikkusega, väärtus vastab %s bitisele võtmele, lubatud on 1024, 2048 ja 4096 bitised võtmed',
             'mac',
-            new Buffer(value, 'hex').length * 8
+            Buffer.from(value, 'hex').length * 8
         );
     }
 
