@@ -505,12 +505,12 @@ IPizza.prototype.validateSignature = function(callback) {
             success: !!success,
             errors: !success
                 ? [
-                    {
-                        field: 'VK_MAC',
-                        error: util.format('Allkirja parameetri %s valideerimine ebaõnnestus.', 'VK_MAC'),
-                        download: true
-                    }
-                ]
+                      {
+                          field: 'VK_MAC',
+                          error: util.format('Allkirja parameetri %s valideerimine ebaõnnestus.', 'VK_MAC'),
+                          download: true
+                      }
+                  ]
                 : false,
             warnings: false
         });
@@ -1160,11 +1160,11 @@ IPizzaValidator.prototype.validate_VK_MAC = function() {
         return util.format('Allkirja parameeter %s peab olema BASE64 formaadis', 'VK_MAC');
     }
 
-    if (new Buffer(value, 'base64').length % 128) {
+    if (Buffer.from(value, 'base64').length % 128) {
         return util.format(
             'Allkirja parameeter %s on vale pikkusega, väärtus vastab %s bitisele võtmele, lubatud on 1024, 2048 ja 4096 bitised võtmed',
             'VK_MAC',
-            new Buffer(value, 'base64').length * 8
+            Buffer.from(value, 'base64').length * 8
         );
     }
 
@@ -1355,9 +1355,9 @@ IPizza.generateForm = function(payment, project, callback) {
                                 url,
                                 fields: payload
                                     ? Object.keys(fields).map(key => ({
-                                        key,
-                                        value: fields[key]
-                                    }))
+                                          key,
+                                          value: fields[key]
+                                      }))
                                     : false
                             };
                         } else {
