@@ -15,7 +15,7 @@ function bodyParser(req, res, next) {
     if (req.method === 'GET' && req.originalUrl.indexOf('?') >= 0) {
         req.body = {};
         req.encodedBody = {};
-        req.rawBody = new Buffer(urllib.parse(req.originalUrl, false, false).query || '', 'utf-8');
+        req.rawBody = Buffer.from(urllib.parse(req.originalUrl, false, false).query || '', 'utf-8');
         return parseUrlencodedBody(req, res, next);
     }
 
@@ -50,7 +50,7 @@ function bodyParser(req, res, next) {
                     break;
                 case 'application/json':
                     try {
-                        req.body = JSON.parse(new Buffer(req.rawBody, 'utf-8'));
+                        req.body = JSON.parse(req.rawBody.toString());
                     } catch (E) {
                         // ignore
                     }
@@ -78,7 +78,7 @@ function parseUrlencodedBody(req, res, next) {
             pos = 0;
 
         if (encodedChars > 0) {
-            buf = new Buffer(len);
+            buf = Buffer.alloc(len);
             for (let i = 0, l = value.length; i < l; i++) {
                 if (value[i] === '%' && value.substr(i, 3).match(/^%[0-9a-fA-F]{2}$/)) {
                     buf[pos] = parseInt(value.substr(i + 1, 2), 16);
@@ -91,7 +91,7 @@ function parseUrlencodedBody(req, res, next) {
             }
             value = buf;
         } else {
-            value = new Buffer(value, 'binary');
+            value = Buffer.from(value, 'binary');
         }
 
         if (key && value && value.length) {
@@ -130,7 +130,7 @@ function parseMultipartBody(req, res, boundary, next) {
         if (headers['content-disposition'] && headers['content-disposition'].value === 'form-data') {
             key = headers['content-disposition'].name;
         }
-        value = new Buffer(he.decode(contents || ''), 'binary');
+        value = Buffer.from(he.decode(contents || ''), 'binary');
 
         if (key && value && value.length) {
             req.body[key] = value.toString('utf-8');
