@@ -25,7 +25,7 @@ module.exports = function(req, res, next) {
             write.call(res, chunk, encoding);
         } else {
             if (typeof chunk === 'string') {
-                chunk = new Buffer(chunk, encoding || 'utf-8');
+                chunk = Buffer.from(chunk, encoding || 'utf-8');
             }
             buffer.push(chunk);
             bufferLen += chunk.length;
@@ -40,7 +40,7 @@ module.exports = function(req, res, next) {
         }
 
         if (buffer) {
-            let buf = bufferLen ? Buffer.concat(buffer, bufferLen) : new Buffer(0);
+            let buf = bufferLen ? Buffer.concat(buffer, bufferLen) : Buffer.from(0);
             let out = encodinglib.convert(buf, res.forceCharset);
 
             res.setHeader('Content-Length', out.length);
