@@ -276,7 +276,7 @@ function initializeAccountTicket(req, address, description, role, callback) {
                 created: new Date()
             };
 
-            db.database.collection('tickets').insert(ticket, err => {
+            db.database.collection('tickets').insertOne(ticket, err => {
                 if (err) {
                     return callback(err);
                 }
