@@ -6,14 +6,24 @@ const MongoClient = require('mongodb').MongoClient;
 let dbObj;
 
 module.exports.init = function(callback) {
-    MongoClient.connect(config.mongodb.url, {}, (err, db) => {
-        if (err) {
-            return callback(err);
+    MongoClient.connect(
+        config.mongodb.url,
+        {
+            useNewUrlParser: true,
+            reconnectTries: 100000,
+            reconnectInterval: 1000
+        },
+        (err, db) => {
+            if (err) {
+                return callback(err);
+            }
+            if (db.s && db.s.options && db.s.options.dbName) {
+                db = db.db(db.s.options.dbName);
+            }
+            module.exports.database = dbObj = db;
+            return callback(null, db);
         }
-
-        module.exports.database = dbObj = db;
-        return callback(null, db);
-    });
+    );
 };
 
 module.exports.save = function(collectionName, record, callback) {
@@ -26,15 +36,20 @@ module.exports.save = function(collectionName, record, callback) {
         }
 
         if (id) {
-            collection.findAndModify(
+            let data = {};
+            Object.keys(record || {}).forEach(key => {
+                if (!['_id'].includes(key)) {
+                    data[key] = record[key];
+                }
+            });
+            collection.findOneAndUpdate(
                 {
                     _id: id
                 },
-                false,
-                record,
+                { $set: data },
                 {
                     upsert: true, // insert as new if not found
-                    new: true // return updated record
+                    returnOriginal: false // return updated record
                 },
                 (err, record) => {
                     if (err) {
@@ -81,13 +96,12 @@ module.exports.modify = function(collectionName, query, update, callback) {
             return callback(err);
         }
 
-        collection.findAndModify(
+        collection.findOneAndUpdate(
             query,
-            false,
             update,
             {
                 upsert: true, // insert as new if not found
-                new: true // return updated record
+                returnOriginal: false // return updated record
             },
             (err, record) => {
                 if (err) {
@@ -129,11 +143,15 @@ module.exports.find = function(collectionName, query, fields, options, callback)
         options = options || {};
         fields = fields || {};
 
+        if (fields) {
+            options.projection = fields;
+        }
+
         if (!('limit' in options)) {
             options.limit = 1000;
         }
 
-        collection.find(query, fields, options).toArray((err, docs) => {
+        collection.find(query, options).toArray((err, docs) => {
             if (err) {
                 return callback(err);
             }
