@@ -27,7 +27,9 @@ function sendInvitation(req, ticket) {
             link: `${req.siteProto}://${req.siteHostname}/account/join?username=${encodeURIComponent(ticket.address)}&ticket=${ticket._id.toString()}`,
             title: 'Loo uus konto'
         },
-        footer: `Said selle e-kirja, kuna saidi <a href="${req.siteProto}://${req.siteHostname}/">${req.siteHostname}</a> administraator soovis lisada Sinu e-posti aadressile uut kasutajakontot`
+        footer: `Said selle e-kirja, kuna saidi <a href="${req.siteProto}://${req.siteHostname}/">${
+            req.siteHostname
+        }</a> administraator soovis lisada Sinu e-posti aadressile uut kasutajakontot`
     });
 
     let mailOptions = {
@@ -58,7 +60,9 @@ function sendRegistration(req, user) {
         title: req.siteTitle,
         message: `${user.name}, sinu ${req.siteTitle} konto on valmis!
 Sinu kasutajanimi: ${user.username}<br/>Sinu parool: registreerimisel kasutatud parool`,
-        footer: `Said selle e-kirja, kuna keegi kasutas sinu aadressi konto loomiseks saidil <a href="${req.siteProto}://${req.siteHostname}/">${req.siteHostname}</a>`
+        footer: `Said selle e-kirja, kuna keegi kasutas sinu aadressi konto loomiseks saidil <a href="${req.siteProto}://${req.siteHostname}/">${
+            req.siteHostname
+        }</a>`
     });
 
     let mailOptions = {
@@ -92,7 +96,9 @@ function sendResetLink(req, user, resetToken) {
             link: `${req.siteProto}://${req.siteHostname}/account/reset-password?username=${encodeURIComponent(user.username)}&resetToken=${resetToken}`,
             title: 'Lähtesta Parool'
         },
-        footer: `Said selle e-kirja, kuna keegi üritas lähtestada sinu konto parooli saidil <a href="${req.siteProto}://${req.siteHostname}/">${req.siteHostname}</a>`
+        footer: `Said selle e-kirja, kuna keegi üritas lähtestada sinu konto parooli saidil <a href="${req.siteProto}://${req.siteHostname}/">${
+            req.siteHostname
+        }</a>`
     });
 
     let mailOptions = {
